Lots of GUI work and refactors. :)

+++ CoFHCore/src/main/java/cofh/core/gui/element/ElementBase.java
@@ -217,7 +217,7 @@ public abstract class ElementBase {
 
     }
 
-    public boolean mouseWheel(double mouseX, double mouseY, int movement) {
+    public boolean mouseWheel(double mouseX, double mouseY, double movement) {
 
         return false;
     }
